zeromq: quick fix to make sub always call setup_socket

+++ gr-zeromq/lib/base_impl.h
@@ -39,12 +39,12 @@ namespace gr {
       std::string endpoint();
 
     protected:
-      base_impl(size_t type, size_t itemsize, size_t vlen, int timeout, bool pass_tags);
+      base_impl(int type, size_t itemsize, size_t vlen, int timeout, bool pass_tags);
       void set_hwm();
       virtual void setup_socket() {};
       zmq::context_t  *d_context;
       zmq::socket_t   *d_socket;
-      size_t          d_type;
+      int             d_type;
       size_t          d_vsize;
       int             d_timeout;
       bool            d_pass_tags;
@@ -62,7 +62,7 @@ namespace gr {
       void set_endpoint(const char* address);
 
     protected:
-      base_sink_impl(size_t type, size_t itemsize, size_t vlen, char *address, int timeout, bool pass_tags, int hwm);
+      base_sink_impl(int type, size_t itemsize, size_t vlen, char *address, int timeout, bool pass_tags, int hwm);
       int send_message(const void *in_buf, const int in_nitems, const uint64_t in_offset);
     };
